API利用サービスアクションを共通化した
