+++ components/screen/form/service.vue
@@ -13,8 +13,13 @@ const emit = defineEmits<{
 }>()
 
 const modal = ref(false)
-const { handleSubmit, handleReset, formData, resetServiceForm } =
-  useServiceForm()
+const {
+  handleSubmit,
+  handleReset,
+  formData,
+  resetServiceForm,
+  changeImageFile,
+} = useServiceForm()
 
 watch(modal, (current) => {
   if (current) {
@@ -22,11 +27,6 @@ watch(modal, (current) => {
     resetServiceForm(props.serviceData)
   }
 })
-
-const onChangeImageFile = async (params: { file: File; url: string }) => {
-  formData.image.value.value = params.url
-  formData.imageFile.value.value = params.file
-}
 
 const onCreate = handleSubmit((serviceForm) => {
   emit('create', serviceForm)
@@ -74,7 +74,7 @@ const onCancel = () => {
           :image-url="formData.image.value.value"
           :error-messages="formData.image.errorMessage.value"
           label="タイトル画像"
-          @change-image-file="onChangeImageFile"
+          @change-image-file="changeImageFile"
         />
       </div>
       <div class="mt-3">
