+++ components/screen/home/index.vue
@@ -61,18 +61,8 @@ const homeSections: DefineHomeSections[] = [
 </template>
 
 <style lang="scss" scoped>
-.section-title {
-  margin-top: 0.5rem;
-  margin-bottom: 0.5rem;
-  text-align: center;
-  span {
-    background: linear-gradient(transparent 75%, orange 75%);
-    font-weight: bold;
-  }
-}
-
 .section-margin {
   padding-top: calc($nav-header-height + 3rem);
-  margin-bottom: 4rem;
+  margin-bottom: 3rem;
 }
 </style>
