+++ components/screen/news/index.vue
@@ -1,0 +1,14 @@
+<template>
+  <article id="news-index-position-top">
+    <section>
+      <ScreenHomeType1Newses />
+    </section>
+  </article>
+</template>
+
+<style lang="scss" scoped>
+#news-index-position-top {
+  padding-top: calc($nav-header-height + 3rem);
+  margin-bottom: 3rem;
+}
+</style>
