+++ pages/news/[id].vue
@@ -1,0 +1,4 @@
+<template>
+  <div>ID</div>
+  <ScreenNews />
+</template>
