+++ pages/news/index.vue
@@ -1,0 +1,4 @@
+<template>
+  <div>INDEX</div>
+  <ScreenNews />
+</template>
